fix(types): make mypy and bandit pass, and stop CI ignoring them
mypy was 103 errors across 26 files and the workflow ran it with
continue-on-error, so nothing it found was ever seen. Deleting the legacy
generator took it to 32; the rest are fixed here and the step now fails the
build. bandit exited 1 on 21 low findings and its step has no
continue-on-error either, so the security job was red too.

Three of the type errors were real defects rather than missing annotations:

`acceptance.step_mismatches` crashed. An expectation naming neither a
`value` nor a `baseline` reached `_satisfied(compare, actual, None)`, where
`at_least` evaluates `actual >= None` and raises TypeError. Nothing in
`services.runtime_test` catches it, so one malformed rule ended a whole run
instead of failing its own step. It is now a mismatch with a sentence saying
what is missing, tested three ways -- including that a stated value of 0 is
still a value, which a falsy check would have got wrong.

`cli._project_command` bound one name to a `SourceResult` in the scaffold
branch and a `BuildResult` in the build branch, then read `.artifact` and
`.success` off it; and bound `dossier` to a non-optional reference in one
branch and to `load_reference`'s optional answer in another. Both now use
distinct names.

`services.add_asset` took `kind: str` and passed it to a field typed
`Literal["sprite", "tileset", "font", "screen"]`, so a typo was caught by
pydantic one call later and by nothing before it. The literal is now a named
`AssetKind` in models.py and the parameter uses it.

The rest are annotations, and two of them close a hole worth naming: the
tile-drawing path typed its source and its artist as `object`, so nothing
checked that either has the methods the loop calls. `TileSource` is now a
Protocol beside `SheetSource`, and `services.draw_tiles` takes a
`TileArtist`.

bandit's findings are all low and all inherent: B404/B603 are this project
running zcc, sdcc, make and an emulator, always with an argument list rather
than a shell string, and B105 fires twelve times on path tokens and gate
report keys. They are skipped in pyproject with the reasoning written down;
anything at medium or high still fails.

Also: CI checks `tests` and `scripts` alongside `llmz80` now, stops
exporting an OpenAI key to a suite that cannot use one, and drops the
`safety check` step -- a deprecated command behind continue-on-error, which
is to say a step that reported nothing whatever it found.

+++ llmz80/quality/emulator_smoke.py
@@ -578,17 +578,20 @@ def scripted_run_seconds(
     screen_cost = exchange if steps and profile["reads_display_file"] else 0.0
     # Rounded up, not truncated: `int` shaved off up to a second from a figure
     # whose entire purpose is not being short.
-    return math.ceil(
-        max(6, seconds)
-        # What it costs to get the program running at all, which is nothing on
-        # a Spectrum -- its .tap autostarts -- and most of a CPC run's first
-        # quarter minute, because a .dsk has to be asked for at a BASIC prompt.
-        + profile["boot_seconds"]
-        + probe_cost
-        + hold_cost
-        + command_cost
-        + screen_cost
-        + _HARNESS_OVERHEAD
+    return int(
+        math.ceil(
+            max(6, seconds)
+            # What it costs to get the program running at all, which is nothing
+            # on a Spectrum -- its .tap autostarts -- and most of a CPC run's
+            # first quarter minute, because a .dsk has to be asked for at a
+            # BASIC prompt.
+            + profile["boot_seconds"]
+            + probe_cost
+            + hold_cost
+            + command_cost
+            + screen_cost
+            + _HARNESS_OVERHEAD
+        )
     )
